Reduce build times by refactoring JSValueInWrappedObject.h https://bugs.webkit.org/show_bug.cgi?id=253098 rdar://106045427
Reviewed by NOBODY (OOPS!).

Separate out the implementation of JSValueInWrappedObject which requires pulling in JSCJSValueInlines.h,
WeakInlines.h, JSDOMWrapaper.h and DOMWrapperWorld.h into a JSValueInWrappedObjectInlines.h header.

This reduces the number of times JSCJSValueInlines.h is included in the WebCore project from 262 to 242
(-20), and JSDOMGlobalObject.h from 243 to 222 (-21).

* Source/WebCore/Modules/indexeddb/IDBCursor.cpp:
* Source/WebCore/Modules/indexeddb/IDBRequest.cpp:
* Source/WebCore/Modules/paymentrequest/PaymentMethodChangeEvent.cpp:
* Source/WebCore/Modules/paymentrequest/PaymentResponse.cpp:
* Source/WebCore/Modules/webaudio/AudioBuffer.cpp:
* Source/WebCore/Modules/webaudio/AudioBuffer.h:
* Source/WebCore/Modules/webaudio/AudioWorkletProcessor.cpp:
* Source/WebCore/Modules/webxr/WebXRRigidTransform.cpp:
* Source/WebCore/Modules/webxr/WebXRView.cpp:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/bindings/js/JSAbortSignalCustom.cpp:
* Source/WebCore/bindings/js/JSDOMConvertAny.h:
* Source/WebCore/bindings/js/JSHistoryCustom.cpp:
* Source/WebCore/bindings/js/JSValueInWrappedObject.h:
(WebCore::JSValueInWrappedObject::JSValueInWrappedObject): Deleted.
(WebCore::JSValueInWrappedObject::getValue const): Deleted.
(WebCore::JSValueInWrappedObject::operator bool const): Deleted.
(WebCore::JSValueInWrappedObject::visit const): Deleted.
(WebCore::JSValueInWrappedObject::setWeakly): Deleted.
(WebCore::JSValueInWrappedObject::set): Deleted.
(WebCore::JSValueInWrappedObject::clear): Deleted.
(WebCore::cachedPropertyValue): Deleted.
* Source/WebCore/bindings/js/JSValueInWrappedObjectInlines.h: Added.
(WebCore::JSValueInWrappedObject::JSValueInWrappedObject):
(WebCore::JSValueInWrappedObject::getValue const):
(WebCore::JSValueInWrappedObject::operator bool const):
(WebCore::JSValueInWrappedObject::visit const):
(WebCore::JSValueInWrappedObject::setWeakly):
(WebCore::JSValueInWrappedObject::set):
(WebCore::JSValueInWrappedObject::clear):
(WebCore::cachedPropertyValue):
* Source/WebCore/bindings/js/JSWebXRRigidTransformCustom.cpp:
* Source/WebCore/dom/AbortSignal.cpp:
* Source/WebCore/dom/AbortSignal.h:
* Source/WebCore/dom/CustomEvent.cpp:
* Source/WebCore/dom/ErrorEvent.cpp:
* Source/WebCore/dom/MessageEvent.cpp:
* Source/WebCore/dom/PopStateEvent.cpp:
* Source/WebCore/dom/PromiseRejectionEvent.cpp:
* Source/WebCore/html/HTMLPlugInElement.cpp:
* Source/WebCore/page/History.cpp:
