+++ Source/WebCore/Modules/indexeddb/IDBCursor.cpp
@@ -34,6 +34,7 @@
 #include "IDBObjectStore.h"
 #include "IDBRequest.h"
 #include "IDBTransaction.h"
+#include "JSValueInWrappedObjectInlines.h"
 #include "Logging.h"
 #include "ScriptExecutionContext.h"
 #include "SerializedScriptValue.h"
