+++ Source/WebCore/Modules/paymentrequest/PaymentMethodChangeEvent.cpp
@@ -28,6 +28,7 @@
 
 #if ENABLE(PAYMENT_REQUEST)
 
+#include "JSValueInWrappedObjectInlines.h"
 #include <wtf/IsoMallocInlines.h>
 
 namespace WebCore {
