+++ Source/WebCore/bindings/js/JSAbortSignalCustom.cpp
@@ -26,6 +26,7 @@
 #include "config.h"
 #include "JSAbortSignal.h"
 
+#include "JSValueInWrappedObjectInlines.h"
 #include "WebCoreOpaqueRootInlines.h"
 
 namespace WebCore {
