+++ Source/WebCore/bindings/js/JSValueInWrappedObject.h
@@ -24,14 +24,21 @@
 
 #pragma once
 
-#include "DOMWrapperWorld.h"
-#include "JSDOMWrapper.h"
-#include <JavaScriptCore/JSCJSValueInlines.h>
-#include <JavaScriptCore/SlotVisitor.h>
-#include <JavaScriptCore/WeakInlines.h>
-#include <variant>
+#include <JavaScriptCore/JSCJSValue.h>
+#include <JavaScriptCore/Weak.h>
+
+namespace JSC {
+
+class JSCell;
+class JSGlobalObject;
+class ThrowScope;
+class VM;
+
+}
 
 namespace WebCore {
+
+class JSDOMObject;
 
 // This class includes a lot of GC related subtle things, and changing this class easily causes GC crashes.
 // Any changes on this class must be reviewed by JavaScriptCore reviewers too.
@@ -40,15 +47,16 @@ class JSValueInWrappedObject {
     WTF_MAKE_NONCOPYABLE(JSValueInWrappedObject);
     WTF_MAKE_NONMOVABLE(JSValueInWrappedObject);
 public:
-    JSValueInWrappedObject(JSC::JSValue = { });
+    inline JSValueInWrappedObject();
+    inline JSValueInWrappedObject(JSC::JSValue);
 
-    explicit operator bool() const;
-    template<typename Visitor> void visit(Visitor&) const;
-    void clear();
+    inline explicit operator bool() const;
+    template<typename Visitor> inline void visit(Visitor&) const;
+    inline void clear();
 
-    void set(JSC::VM&, const JSC::JSCell* owner, JSC::JSValue);
-    void setWeakly(JSC::JSValue);
-    JSC::JSValue getValue(JSC::JSValue nullValue = JSC::jsUndefined()) const;
+    inline void set(JSC::VM&, const JSC::JSCell* owner, JSC::JSValue);
+    inline void setWeakly(JSC::JSValue);
+    inline JSC::JSValue getValue(JSC::JSValue nullValue = JSC::jsUndefined()) const;
 
 private:
     // Keep in mind that all of these fields are accessed concurrently without lock from concurrent GC thread.
@@ -56,70 +64,6 @@ private:
     JSC::Weak<JSC::JSCell> m_cell { };
 };
 
-JSC::JSValue cachedPropertyValue(JSC::ThrowScope&, JSC::JSGlobalObject&, const JSDOMObject& owner, JSValueInWrappedObject& cacheSlot, const Function<JSC::JSValue(JSC::ThrowScope&)>&);
-
-inline JSValueInWrappedObject::JSValueInWrappedObject(JSC::JSValue value)
-{
-    setWeakly(value);
-}
-
-inline JSC::JSValue JSValueInWrappedObject::getValue(JSC::JSValue nullValue) const
-{
-    if (m_nonCell)
-        return m_nonCell;
-    return m_cell ? m_cell.get() : nullValue;
-}
-
-inline JSValueInWrappedObject::operator bool() const
-{
-    return m_nonCell || m_cell;
-}
-
-template<typename Visitor>
-inline void JSValueInWrappedObject::visit(Visitor& visitor) const
-{
-    visitor.append(m_cell);
-}
-
-template void JSValueInWrappedObject::visit(JSC::AbstractSlotVisitor&) const;
-template void JSValueInWrappedObject::visit(JSC::SlotVisitor&) const;
-
-inline void JSValueInWrappedObject::setWeakly(JSC::JSValue value)
-{
-    if (!value.isCell()) {
-        m_nonCell = value;
-        m_cell.clear();
-        return;
-    }
-    m_nonCell = { };
-    JSC::Weak weak { value.asCell() };
-    WTF::storeStoreFence();
-    m_cell = WTFMove(weak);
-}
-
-inline void JSValueInWrappedObject::set(JSC::VM& vm, const JSC::JSCell* owner, JSC::JSValue value)
-{
-    setWeakly(value);
-    vm.writeBarrier(owner, value);
-}
-
-inline void JSValueInWrappedObject::clear()
-{
-    m_nonCell = { };
-    m_cell.clear();
-}
-
-inline JSC::JSValue cachedPropertyValue(JSC::ThrowScope& throwScope, JSC::JSGlobalObject& lexicalGlobalObject, const JSDOMObject& owner, JSValueInWrappedObject& cachedValue, const Function<JSC::JSValue(JSC::ThrowScope&)>& function)
-{
-    if (cachedValue && isWorldCompatible(lexicalGlobalObject, cachedValue.getValue()))
-        return cachedValue.getValue();
-
-    auto value = function(throwScope);
-    RETURN_IF_EXCEPTION(throwScope, { });
-
-    cachedValue.set(lexicalGlobalObject.vm(), &owner, cloneAcrossWorlds(lexicalGlobalObject, owner, value));
-    ASSERT(isWorldCompatible(lexicalGlobalObject, cachedValue.getValue()));
-    return cachedValue.getValue();
-}
+inline JSC::JSValue cachedPropertyValue(JSC::ThrowScope&, JSC::JSGlobalObject&, const JSDOMObject& owner, JSValueInWrappedObject& cacheSlot, const Function<JSC::JSValue(JSC::ThrowScope&)>&);
 
 } // namespace WebCore
