+++ Source/WebCore/bindings/js/JSWebXRViewCustom.cpp
@@ -29,6 +29,7 @@
 #include "JSWebXRView.h"
 
 #include "JSDOMConvertBufferSource.h"
+#include "JSValueInWrappedObjectInlines.h"
 #include "WebXRView.h"
 
 namespace WebCore {
