+++ Source/WebCore/dom/PromiseRejectionEvent.cpp
@@ -28,6 +28,7 @@
 
 #include "DOMWrapperWorld.h"
 #include "JSDOMPromise.h"
+#include "JSValueInWrappedObjectInlines.h"
 #include <JavaScriptCore/HeapInlines.h>
 #include <JavaScriptCore/StrongInlines.h>
 #include <wtf/IsoMallocInlines.h>
